+++ app/src/main/java/com/burhanrashid52/photoeditor/ColorPickerAdapter.java
@@ -1,4 +1,4 @@
-package com.burhanrashid52.imageeditor;
+package com.burhanrashid52.photoeditor;
 
 import android.content.Context;
 import android.graphics.Color;
