+++ app/src/main/java/com/burhanrashid52/photoeditor/PropertiesBSFragment.java
@@ -1,4 +1,4 @@
-package com.burhanrashid52.imageeditor;
+package com.burhanrashid52.photoeditor;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
