+++ app/src/main/java/com/burhanrashid52/photoeditor/StickerBSFragment.java
@@ -1,8 +1,12 @@
-package com.burhanrashid52.imageeditor;
+package com.burhanrashid52.photoeditor;
 
 import android.annotation.SuppressLint;
 import android.app.Dialog;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.design.widget.BottomSheetBehavior;
 import android.support.design.widget.BottomSheetDialogFragment;
 import android.support.design.widget.CoordinatorLayout;
@@ -11,22 +15,22 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+import android.widget.ImageView;
 
-import java.util.ArrayList;
+public class StickerBSFragment extends BottomSheetDialogFragment {
 
-import ja.burhanrashid52.photoeditor.PhotoEditor;
-
-public class EmojiBSFragment extends BottomSheetDialogFragment {
-
-    public EmojiBSFragment() {
+    public StickerBSFragment() {
         // Required empty public constructor
     }
 
-    private EmojiListener mEmojiListener;
+    private StickerListener mStickerListener;
 
-    public interface EmojiListener {
-        void onEmojiClick(String emojiUnicode);
+    public void setStickerListener(StickerListener stickerListener) {
+        mStickerListener = stickerListener;
+    }
+
+    public interface StickerListener {
+        void onStickerClick(Bitmap bitmap);
     }
 
     private BottomSheetBehavior.BottomSheetCallback mBottomSheetBehaviorCallback = new BottomSheetBehavior.BottomSheetCallback() {
@@ -44,6 +48,7 @@ public class EmojiBSFragment extends BottomSheetDialogFragment {
         }
     };
 
+
     @SuppressLint("RestrictedApi")
     @Override
     public void setupDialog(Dialog dialog, int style) {
@@ -59,54 +64,72 @@ public class EmojiBSFragment extends BottomSheetDialogFragment {
         ((View) contentView.getParent()).setBackgroundColor(getResources().getColor(android.R.color.transparent));
         RecyclerView rvEmoji = contentView.findViewById(R.id.rvEmoji);
 
-        GridLayoutManager gridLayoutManager = new GridLayoutManager(getActivity(), 5);
+        GridLayoutManager gridLayoutManager = new GridLayoutManager(getActivity(), 3);
         rvEmoji.setLayoutManager(gridLayoutManager);
-        EmojiAdapter emojiAdapter = new EmojiAdapter();
-        rvEmoji.setAdapter(emojiAdapter);
+        StickerAdapter stickerAdapter = new StickerAdapter();
+        rvEmoji.setAdapter(stickerAdapter);
     }
 
-    public void setEmojiListener(EmojiListener emojiListener) {
-        mEmojiListener = emojiListener;
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
     }
 
+    public class StickerAdapter extends RecyclerView.Adapter<StickerAdapter.ViewHolder> {
 
-    public class EmojiAdapter extends RecyclerView.Adapter<EmojiAdapter.ViewHolder> {
-
-        ArrayList<String> emojisList = PhotoEditor.getEmojis(getActivity());
+        int[] stickerList = new int[]{R.drawable.aa, R.drawable.bb};
 
         @Override
         public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-            View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.row_emoji, parent, false);
+            View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.row_sticker, parent, false);
             return new ViewHolder(view);
         }
 
         @Override
         public void onBindViewHolder(ViewHolder holder, int position) {
-            holder.txtEmoji.setText(emojisList.get(position));
+            holder.imgSticker.setImageResource(stickerList[position]);
         }
 
         @Override
         public int getItemCount() {
-            return emojisList.size();
+            return stickerList.length;
         }
 
         class ViewHolder extends RecyclerView.ViewHolder {
-            TextView txtEmoji;
+            ImageView imgSticker;
 
             ViewHolder(View itemView) {
                 super(itemView);
-                txtEmoji = itemView.findViewById(R.id.txtEmoji);
+                imgSticker = itemView.findViewById(R.id.imgSticker);
 
                 itemView.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
-                        if (mEmojiListener != null) {
-                            mEmojiListener.onEmojiClick(emojisList.get(getLayoutPosition()));
+                        if (mStickerListener != null) {
+                            mStickerListener.onStickerClick(
+                                    BitmapFactory.decodeResource(getResources(),
+                                            stickerList[getLayoutPosition()]));
                         }
                         dismiss();
                     }
                 });
             }
         }
+    }
+
+    private String convertEmoji(String emoji) {
+        String returnedEmoji = "";
+        try {
+            int convertEmojiToInt = Integer.parseInt(emoji.substring(2), 16);
+            returnedEmoji = getEmojiByUnicode(convertEmojiToInt);
+        } catch (NumberFormatException e) {
+            returnedEmoji = "";
+        }
+        return returnedEmoji;
+    }
+
+    private String getEmojiByUnicode(int unicode) {
+        return new String(Character.toChars(unicode));
     }
 }
