+++ app/src/main/java/com/burhanrashid52/photoeditor/base/BaseActivity.java
@@ -1,4 +1,4 @@
-package com.burhanrashid52.imageeditor.base;
+package com.burhanrashid52.photoeditor.base;
 
 import android.app.ProgressDialog;
 import android.content.pm.PackageManager;
