+++ app/src/main/java/com/burhanrashid52/photoeditor/base/BaseFragment.java
@@ -1,4 +1,4 @@
-package com.burhanrashid52.imageeditor.base;
+package com.burhanrashid52.photoeditor.base;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
