+++ app/src/main/java/com/burhanrashid52/photoeditor/tools/ToolType.java
@@ -1,4 +1,4 @@
-package com.burhanrashid52.imageeditor.tools;
+package com.burhanrashid52.photoeditor.tools;
 
 /**
  * @author <a href="https://github.com/burhanrashid52">Burhanuddin Rashid</a>
