+++ app/src/test/java/com/burhanrashid52/photoeditor/SampleTest.java
@@ -1,4 +1,4 @@
-package com.burhanrashid52.imageeditor;
+package com.burhanrashid52.photoeditor;
 
 import org.junit.Test;
